Add function for exercice 04
I had a bit of trouble with dealing with an undefined number of
arguments. Si I used the rest parameter syntax and not the arguments
object syntax, because I wanted all of the other arguments apart from
the input Array. Then I fugured out a way to just add the elements from
the inputArray which didn't match any of the arguments, so I used a
label for the first loop to continue within the second loop so I don't
write in the output if I ever encounter the same argument. Also I
realised that it was better to use a for...of loop because all I wanted
was to scroll trhough an array.