Create the first draft of the function
It passes the 6 first tests. I used an array to store every number of
the finonacci suite. I initiated it with 0 and 1 so it can return
directly the index of the array with the number input (without the gap).
Then a for loop creates the suite the size of the number input.